v0.3-spring

+++ src/main/java/com/api_integrador_utp/repository/ProductoRepository.java
@@ -4,10 +4,11 @@ import com.api_integrador_utp.entity.Producto;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface ProductoRepository extends JpaRepository<Producto, Integer> {
+public interface ProductoRepository extends JpaRepository<Producto, Long> {
     Optional<Producto> findByNombre(String nombre);
     boolean existsByNombre(String nombre);
 }
